requirements.txt
feedparser>=6.0.0
requests>=2.28.0
beautifulsoup4>=4.12.0
lxml>=4.9.0
PyYAML>=6.0
jinja2>=3.1.0
schedule>=1.2.0
